Add files via upload
#Fundamentals of Computer Programming Lab task

In this task, you have to understand and implement the following concepts
- Variables
- IF-ELSE conditioning

##Objectives
• To develop skills for using if – else statements.
• To understand how to program conditional calculations.

##Tools/Software Requirement
-Microsoft Visual Studio 2010 or later

##Problem Statement:
In this assignment you will create a program that an employee can use to calculate his/her
pension. The pension is calculated by applying some calculations on his last drawn salary.

**Details about salary:**
Employee's salary is divided into three parts. One part is the basic salary, the second part is the
house rent and the third part is the old age allowance. These three combine to form an employees
total salary.

5% income tax and 7% provincial tax is deducted from the basic pay.
The house rent and old age allowance are calculated after the deduction of tax.
Employee who is less than 45 years of age, gets no old age allowance.
For an employee who is between 45 and 55, old age allowance is 10%.
For employees older than 55 the old age allowance is 15%
Married employees are not given any house rent and the married get 15%

**Pension calculation:**
An employee's pension is calculated according to the following rules:
1. Basic pay is doubled and multiplied with the number of months is service.
2. House rent is multiplied with the months in service since marriage.
3. Old age allowance is multiplied with 3.
4. Total pay is multiplied with 2.
5. The total of the above four are added together to compute the pension.
Take necessary input from a user and calculate her/his pension. You cannot ask the employee
about the months in service, or the months in service since marriage. Employee's input should be
a date and you need to calculate the months in your program.
The output of the program must be very structured and detailed. It not only shows the pension
but also must show all the calculations under proper heads and step by step so the employee can
understand how the pension is computed.